Added a stub/testing class for pushing individual metrics to RabbitMQ on ingest. It didn't affect the CPU usage much (memory not at all) even while pushing 35K metrics per second through a single TSD. So it looks quite viable for everything else I have in store. I will probably replace it with ZeroMQ so folks don't need to learn another tech to use OpenTSDB. Added Query.getTimeTaken() to stub out tracking query time Modified TSDB.getTimeSeriesMeta() to return null if the meta wasn't found Added some tracing to TsdbQuery while I'm working on it Added TsdbQuery.setTimeSeries() to take a timeseries UID and parse it out into a query. It works for a single TSUID and I have some stub code for handling multiple TSUIDs in the same spanner, but it didn't work the way I intended so I need to revisit it later to see if it's possible. Modified TsdbQuery.findGroupBys() to check for a / character to see if the user wants a regex. This will work with Lucene but I need to figure out how to run regex queries through the index first. Changed time_taken for TsdbJSONStats to a double instead of a float Changed the Booleans to boolean in TimeSeriesMeta Fixed the JSON serialization in TimeSeriesMeta where it couldn't dump a GeneralMeta as the key to a map. Had to convert it back into an array of GeneralMeta objects and now they show up properly. Added group_only and terms to SearchQuery along with parsing code Modified SearchQuery.setTotal_hits() to handle term queries properly Fixed Searcher.groupBy() to handle null values for the group or sub group Added code to Searcher.GroupBy() that will return just the group names if requested. Don't really need it with the Terms method Added Searcher.getTerms() that returns a list of unique values for a specific field in the Lucene index Modified GroupRPC to handle the new terms and group_only queries Modified MetaRPC to handle nulls returned from get meta now Modified QueryHandler query classes to handle the TSUID requests properly and parse everything Restored the CompactHashSet from Ontopia to TimeseriesUID which is actually much less RAM intensive

+++ src/formatters/TsdbJSON.java
@@ -268,7 +268,7 @@ public class TsdbJSON extends TSDFormatter {
     /** The number of datapoints actually returned */
     public long aggregated_datapoints;
     /** The amount of time it took to generate the results */
-    public float time_taken;
+    public double time_taken;
     /** The number of unique timeseries IDs in the results */
     public long unique_tsids;
   }
